requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
qrcode==8.2
Pillow==11.1.0
gunicorn==23.0.0
pytest==8.3.5
python-dotenv==1.1.0
Flask-WTF
email-validator